:rocket: Swap gender in section title of TLJ post
Decided I liked better the subtle layer of twisting the White Christmas lyric to possible reference General Organa, both in the immediate case of the film's weird coma part, and the struggle to replace her in the next episode. It might the fact that it is a song lyric less obvious, by not directly quoting it, but hopefully the added layers make up for that.